requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
color-it==2.1.3
getmac==0.8.2
icmplib==2.0
idna==2.10
mac-vendor-lookup==0.1.11
multidict==5.0.2
pyip==0.7
python-nmap==0.6.1
pythonping==1.0.15
requests==2.25.0
scapy==2.4.4
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
